Move salesforcedx-vscode-apex-replay-debugger tests to vscode-integration dir
